ファイルとコンポーネントの名前を命名規則に従うようにした。
