+++ system/src/App.vue
@@ -11,13 +11,8 @@
 </template>
 
 <script>
-//import HelloWorld from './components/HelloWorld.vue'
-
 export default {
   name: "App",
-  /*components: {
-    HelloWorld
-  }*/
   data() {
     return {
       formListDataObj: null,
